# Optimizations for the `qibo.quantum_info.basis.pauli_basis` and `vectorization` function