--- a/src/qibo/quantum_info/basis.py
+++ b/src/qibo/quantum_info/basis.py
@@ -92,11 +92,7 @@ def pauli_basis(
 
     pauli_labels = {"I": matrices.I, "X": matrices.X, "Y": matrices.Y, "Z": matrices.Z}
     basis_single = backend.cast([pauli_labels[label] for label in pauli_order])
-    einsum = (
-        np.einsum 
-        if backend.name == "tensorflow" 
-        else backend.np.einsum
-    )
+    einsum = np.einsum if backend.name == "tensorflow" else backend.np.einsum
 
     if nqubits > 1:
         dim = 2**nqubits