This improves the construction of the pauli basis by moving everything to tensor notation and removing loops. Namely the `basis_full` is now constructed via contraction through `einsum`. This should also scale well with GPU backends, weirdly for standard numpy CPU there is no speedup.
These are the results:
```
nqubits    old (numpy)    new (numpy)    new (cupy)
4            ~ 1s            ~ 1s          ~ 2.7s
7           ~ 9.5s          ~ 9.1s         ~ 3.5s
8            killed          memory         memory
```
The GPU always takes ~1s to set up. With 8 qubits, `old` gets killed together with the shell session, whereas `new`, both CPU and GPU, raise an out of memory error. To run 8 qubits you apparently need ~64GB of memory.

<s>To perform the einsum I use all the 48 ascii characters available, which means that we are limited to `48/3=16` qubits, unless other characters can be used in einsum. In any case, the memory requirements for 16 qubits are probably going to be very taxing</s> I am using integers as the indices for the `einsum` now, thus this is not limited anymore by the number of ascii charachters. It may be possible to obtain a speedup with numba as well, but I still have to investigate.

EDIT 1: unfortunately `einsum` is not supported by `numba`, however if you `jit` the old implementation:
``` python
            basis_single = list(product(basis_single, repeat=nqubits))

            @njit(fastmath=True, parallel=True, cache=True)
            def _build_basis(basis_single):
                rows = [basis_single[0][0] for i in range(len(basis_single))]
                for i in prange(len(basis_single)):
                    row = basis_single[i][0]
                    for j in range(len(basis_single[i][1:])):
                        row = np.kron(row, basis_single[i][j])
                    rows[i] = row
                return rows

            basis_full = backend.cast(_build_basis(basis_single))
```
<s>you are able to get a nice speedup ~ 5s for 7 qubits, however this happens from the second call, since the first time you are still dealing with the compilation</s> unfortunately this implementation was yielding wrong results, thus I had to rollback to the `einsum` aprroach. Further investigation is needed to understand if it's possible to parallelize and improve this with `numba`.

EDIT 2:
At some point I realized that a possible bottleneck, or better inefficiency, was due to the `qibo.quantum_info.superoperator_transformation.vectorization` that allowed to be run on a single input only, either a state vector or a density matrix, thus forcing to run loops on each element of the basis (which can grow large very quickly). The impact on the runtime is still marginal for the cpu as for 7 qubits its contribute was around ~1-2s out 10s, but for GPUs this starts becoming relevant. Furthermore, `vectorization` appears to be widely used in the `quantum_info` module, which convinced me to generalize it to accept batches of state vectors or density matrices, lifting therefore the need of explicit loops and rather leveraging tensor primitives directly. This was applied to the `pauli_basis` for now but has to be propagated throughout the whole module.

Checklist:
- [ ] Reviewers confirm new code works as expected.
- [x] Tests are passing.
- [x] Coverage does not decrease.
- [x] Documentation is updated.