@alecandido 2024-09-21
@BrunoLiegiBastonLiegi you don't have to use ascii characters for `einsum`

> [`einsum`](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html#numpy.einsum) also provides an alternative way to provide the subscripts and operands as `einsum(op0, sublist0, op1, sublist1, ..., [sublistout])`. If the output shape is not provided in this format [`einsum`](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html#numpy.einsum) will be calculated in implicit mode, otherwise it will be performed explicitly. The examples below have corresponding [`einsum`](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html#numpy.einsum) calls with the two parameter methods.

https://numpy.org/doc/stable/reference/generated/numpy.einsum.html#numpy-einsum

In this way, you can use integers instead of characters, that is much better if you are generating the `einsum` input (and you're not limited by the amount of characters).