@renatomello 2024-09-23
One way we could move forward is to return `generators` for `n >= 8` and let the user do its memory management by accessing the elements on the fly.