@BrunoLiegiBastonLiegi 2024-09-23
> One way we could move forward is to return `generators` for `n >= 8` and let the user do its memory management by accessing the elements on the fly.

mmh yeah, that's a possibility, but only in certain cases. For the sake of parallelization, surely having everything represented by a single tensor contraction is better, but you are limited by memory. 
Maybe, what we could do is defining the generator of a single element of the basis for each backend:
``` python
@cache
def _pauli_basis_element(self, i, nqubits):
    # do things according to backend
```
and then you can build the complete basis as a generator as you suggested 
``` python
def pauli_basis(...):
    return (backend._pauli_basis_element(i, nqubits) for i in range(len))
```
but GPU wise this will be less efficient and it will work, anyway, as long as you don't need the complete basis at the same time.