@BrunoLiegiBastonLiegi 2024-09-24
In the end I was not able to make the `numba` parallelized implementation to work, it was working with `parallel=False` but there was no speed up in that case. Thus I just rolled back to using `einsum` in every case. We are still getting a nice scaling with GPUs. I will get rid of the ascii charachters now.