@BrunoLiegiBastonLiegi 2024-09-25
Ok I ended up also adding a generalization to the `qibo.quantum_info.superoperator_transformation.vectorization`, which now accepts batches of state vectors or density matrices as well, thus allowing to vectorize the whole basis in one shot rather than looping. In the process, I found a weird behaviour of `tensorflow`. Namely, if you have a complex tensor the elements with zero real part, but still non zero imaginary part, are not recovered by `nonzero`:
``` python
import tensorflow as tf
import tensorflow.experimental.numpy as tnp 

a = tf.Variable([0 + 1j, 1 + 0j, 1 + 1j])
tnp.nonzero(a)
# this finds the last two only
# [<tf.Tensor: shape=(2,), dtype=int64, numpy=array([1, 2])>]
```