@renatomello 2024-10-07
@stavros11 since you looked into #1462, could you look into this too? Thanks.